Changed the scalar multiplication to a float like it should have been.

+++ Software-Rendering/src/Vector3f.cpp
@@ -51,7 +51,7 @@ Vector3f Vector3f::operator*(const Vector3f& multVec)
 	return Vector3f(m_x * multVec.m_x , m_y * multVec.m_y , m_z * multVec.m_z);
 }
 //Scalar multiplication.
-Vector3f Vector3f::operator*(const int scalar)
+Vector3f Vector3f::operator*(const float scalar)
 {
 	return Vector3f(scalar * m_x , scalar * m_y, scalar * m_z);
 }
